paid-blocks: refactor move isUpgradable helper fn
