+++ extensions/shared/blocks/cover/index.js
@@ -13,8 +13,8 @@ import { addFilter } from '@wordpress/hooks';
 import coverEditMediaPlaceholder from './cover-media-placeholder';
 import coverMediaReplaceFlow from './cover-replace-control-button';
 import jetpackCoverBlockEdit from './edit';
-import { isUpgradable } from './utils';
 import './editor.scss';
+import { isUpgradable } from "../../paid-blocks/utils";
 
 const addVideoUploadPlanCheck = ( settings, name ) => {
 	if ( ! settings.isDeprecation && isUpgradable( name ) ) {
