+++ extensions/shared/blocks/cover/utils.js
@@ -6,8 +6,6 @@ import { values } from 'lodash';
 /**
  * Internal dependencies
  */
-import { isSimpleSite } from '../../site-type-utils';
-import getJetpackExtensionAvailability from '../../get-jetpack-extension-availability';
 import getAllowedMimeTypesBySite, {
 	getAllowedVideoTypesByType,
 	pickFileExtensionsFromMimeTypes,
@@ -42,21 +40,4 @@ export function isVideoFile( file ) {
 	}
 
 	return false;
-}
-
-/**
- * Check if the cover block should show the upgrade nudge.
- *
- * @param {string} name - Block name.
- * @returns {boolean} True if it should show the nudge. Otherwise, False.
- */
-export function isUpgradable( name ) {
-	const { unavailableReason } = getJetpackExtensionAvailability( 'videopress' );
-
-	return (
-		name &&
-		name === 'core/cover' && // upgrade only for cover block
-		isSimpleSite() && // only for Simple sites
-		[ 'missing_plan', 'unknown' ].includes( unavailableReason )
-	);
 }
