+++ extensions/shared/paid-blocks/utils.js
@@ -12,6 +12,7 @@ import { addQueryArgs } from '@wordpress/url';
  * Internal dependencies
  */
 import { isSimpleSite } from '../site-type-utils';
+import getJetpackExtensionAvailability from '../get-jetpack-extension-availability';
 import getSiteFragment from "../get-site-fragment";
 
 /**
@@ -61,3 +62,22 @@ export function getUpgradeUrl( { planSlug, plan, postId, postType } ) {
 			redirect_to,
 		} );
 }
+
+/**
+ * Check if the cover block should show the upgrade nudge.
+ *
+ * @param {string} name - Block name.
+ * @returns {boolean} True if it should show the nudge. Otherwise, False.
+ */
+export function isUpgradable( name ) {
+	// Remove the namespace form the block name.
+	const blockName = name.replace( /[a-zA-Z_-]+\//, '' );
+	const { unavailableReason } = getJetpackExtensionAvailability( blockName );
+
+	return (
+		name &&
+		isSimpleSite() &&
+		[ 'missing_plan', 'unknown' ].includes( unavailableReason )
+	);
+}
+
